refactor(api-server): split ai.rs into routes/ai/{sessions,equipment,workflows,voice,llm,mod}
Behaviour-preserving module split of the ~3.1k-LOC routes/ai.rs into
surface-scoped sub-modules. Handlers are moved verbatim; every route
registration, authz/RLS guard, and security fix (voice-device-IDOR,
equipment-IDOR, lease/template/photo/generation-request org-scoping,
mandatory voice-OAuth token encryption) is carried over unchanged.

mod.rs re-exports the five public router constructors
(ai_chat_router, sentiment_router, equipment_router, workflow_router,
llm_router) plus the shared MessageResponse/PaginationQuery/AlertsQuery
types, so external references (lib.rs) keep working unchanged. Verified
identical 58-route set vs origin/dev.
